YYC-111: extract memory tests into memory/tests.rs
Acceptance criterion calls for memory/mod.rs to drop below ~500
lines. The inline test module was keeping mod.rs at 762 lines;
moving it to its own file brings mod.rs down to 409.

Tests behavior unchanged.
